fix-9999

+++ Common/Models/Event.cs
@@ -7,7 +7,7 @@ namespace Common.Models
     public class Event
     {
         public int Id { get; set; }
-        public double Total { get; set; }
+        public int Total { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public int AuthorId { get; set; }
